==> _old/code/2-calculate-indicators.ipynb <==
# Calculate indicators

This notebook enables the user to use GDP, population, or other sets of indicators to calculate combined indicators such as per capita emissions or emissions / final energy use.

It should be used on data that is already pre-processed for this toolset to ensure efficiency and avoid errors.

One cell allows the user to adjust the units of the indicator. This should be done with care and the default option is not to do it! 

# import modules

# system
import sys, os, re
import time
import glob

# analytic
import pandas as pd
#import numpy as np

# open cliamte data
#import countrygroups

# plotting
#import seaborn
import matplotlib.pyplot as plt

# global stocktake tools
import gst_tools.gst_utils as utils


from a_parameters import *

variable_name_to_display, proc_data_fname, source_name = utils.get_primap_variable_and_and_file_name(gas_names[raw_entity], raw_sector, raw_scenario, version)
#data_set_1 = 'R-Andrew-2018_cement-CO2.csv'
data_set_1 = proc_data_fname#'PRIMAP-HISTCR_v2.3.1_CH4_total_excl_LULUCF.csv'
#data_set_2 = 'UN-2017-population.csv'
data_set_2, dset2_name = utils.define_dataset(population_fname,gdp_fname,other_fname, population_dset_name, gdp_dset_name, other_dset_name, population,gdp)

# Use this to generate the prefix of the output data file to include the source names of the original data. 
# The entities will automaticall be added when running the script
#new_source_name = 'R-Andrew-2018-cement-CO2-UN-population'
new_source_name = source_name + '_' + dset2_name
with open('gst_tools/combined_source.txt', 'w') as f:
    f.write(new_source_name)
# (don't need the file type ending!)



# get and clean data 

fname_in1 = os.path.join('proc-data', data_set_1)
#EPO#fname_in2 = os.path.join('proc-data', data_set_2)
fname_in2 = os.path.join('gst_tools', 'data', data_set_2)

# read in the data
var1 = pd.read_csv(fname_in1)
# WARNING: The header structure must be specified to the user.
var2 = pd.read_csv(fname_in2, header=2)

# EPO: make country column and variable name columns coincide
var2.rename(columns={'Country Code':'country','Indicator Name':'variable'},inplace=True)

# EPO: Other dataset
if data_set_2 != None:
    if population == True:
        var2_unit = ['Pers']*len(var2)
        var2['unit'] = var2_unit
    elif gdp == True:
        var2_unit = ['USD']*len(var2)
        var2['unit'] = var2_unit
    else:
        var2_unit = [other_dset_unit]*len(var2)
        var2['unit'] = var2_unit
else:
    print('The secondary dataset is not defined. Please check.')

# make sure that the same countries and years are available
var1, var2 = utils.ensure_common_years(var1, var2)
var1, var2 = utils.ensure_common_countries(var1, var2)

# check the data format
check1 = utils.verify_data_format(var1)
check2 = utils.verify_data_format(var2)

if not check1 or not check2:
    print('One of the dataframes is not correct! Please check and try again!')
else:
    # get metadata for later use and checking
    var1_name = var1['variable'].unique()[0]
    var2_name = var2['variable'].unique()[0]

    var1_unit  = var1['unit'].unique()[0]
    var2_unit = var2['unit'].unique()[0]


# combine data...

# for all of these, it's always var1 divided by var 2 and we want to ensure that this is done on countries. 
# Everything else should be consant across the table

def prep_df_for_division(df):
    
    df = df.set_index('country')
    
    year_cols = [y for y in df[df.columns] if (re.match(r"[0-9]{4,7}$", str(y)) is not None)]
    other_cols = list(set(df.columns) - set(year_cols))
    
    df = df.drop(other_cols, axis='columns')
    
    return df
    
# strip original metadata
var1 = prep_df_for_division(var1)
var2 = prep_df_for_division(var2)

# calculate new variables
new_df = var1 / var2

# generate new metadata
new_variable_name = var1_name + '-per-' + var2_name
with open('gst_tools/name_relative_variable.txt', 'w') as f:
    f.write(new_variable_name)
new_df['variable'] = new_variable_name

# automatically generate the unit 
new_df['unit'] = var1_unit + ' / ' + var2_unit
    
new_df = new_df.reset_index()

# reorganise dataframe
new_df = utils.check_column_order(new_df)


# take a look at your new data frame

new_df

# If desired, you can set the unit of the new data here. 

# ****** BE CAREFUL!!! *******
# This option is just to allow you to make the name nicer for plots and will not actually change the units. An 'org_unit' column will 
# be added to teh dataframe for safety. 
# Recommended option: set to False and keep automatic calculation of units.

#convert_unit = False


if convert_unit == False:
    final_df = utils.convert_to_kt(new_df, population, gdp, other_unit=other_dset_unit)
    
else:    
    final_df = new_df

## write the data to file

"""
First ensure that years, unit, 'country', and variable are all in data. If they are
can proceed to print data
"""
# adapted by EPO (change from new_df to final_df)
if 'country' not in final_df.columns or 'unit' not in final_df.columns:
    
    print('Missing required information! Please check your input data and processing!')
    
else:
    
    # define filename as composite of variable and source name
    fname_out = new_source_name + '_' + new_variable_name + '.csv'
    with open('gst_tools/name_relative_dset.txt', 'w') as f:
        f.write(fname_out)
    fullfname_out = os.path.join('proc-data', fname_out)

    # check folder exists
    if not os.path.exists('proc-data'):
        os.makedirs('proc-data')

    # check if file already exists
    files_present = glob.glob(fullfname_out)

    # if no matching files, write to csv, if there are matching files, print statement
    if not files_present:
        # write to csv in proc data folder
        final_df.to_csv(fullfname_out, index=False)
    
    else:
        print('WARNING: This file already exists! adding a date stamp to the file name.')
        fname_out = new_source_name + '_' + new_variable_name + time.strftime("%Y%m%d-%H") + '.csv' 
        final_df.to_csv(fullfname_out, index=False)

    # celebrate success 
    print('Processed data written to file!')
    print(fullfname_out)
